[GSCOLLECT-1115] implement <Primitive>Iterator.remove()
git-svn-id: svn+ssh://gscollections.svn.services.gs.com/svnroot/gscollections-svn/trunk@649 d5c9223b-1aff-41ac-aadd-f810b4a99ac4

+++ unit-tests/src/test/java/com/gs/collections/impl/list/mutable/primitive/AbstractBooleanListTestCase.java
@@ -296,7 +296,7 @@ public abstract class AbstractBooleanListTestCase extends AbstractMutableBoolean
     {
         super.toArray();
         MutableBooleanList list = this.classUnderTest();
-        Assert.assertEquals(3L, list.toArray().length);
+        Assert.assertEquals(3L, (long) list.toArray().length);
         Assert.assertTrue(list.toArray()[0]);
         Assert.assertFalse(list.toArray()[1]);
         Assert.assertTrue(list.toArray()[2]);
